requirements.txt
Flask==2.2.2
tensorflow==2.13.0
Werkzeug==2.2.2
keras==2.13.1
Pillow==9.3.0